pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "numeo-reddit-research"
version = "0.1.0"
description = "Reddit market-research ingestion + RAG for the Numeo growth team"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Proprietary" }
authors = [{ name = "Javohir Akramov" }]
dependencies = [
    "praw>=7.7",
    "anthropic>=0.40",
    "pinecone>=5.0",
    "pyyaml>=6.0",
    "typer>=0.12",
    "python-dotenv>=1.0",
    "rich>=13.7",
]

[project.scripts]
numeo-reddit = "numeo_reddit.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
numeo_reddit = ["py.typed"]
